Split modules, rename tests
